+++ src/euroeval/dataset_configs/__init__.py
@@ -41,6 +41,7 @@ from .slovak import *  # noqa: F403
 from .slovene import *  # noqa: F403
 from .spanish import *  # noqa: F403
 from .swedish import *  # noqa: F403
+from .translation_configs import *  # noqa: F403
 from .ukrainian import *  # noqa: F403